feat(license_data) add license data endpoint for sales data retrieval (#1411)

+++ kong/enterprise_edition/db/migrations/enterprise/004_036-2_to_037.lua
@@ -27,7 +27,7 @@ return {
        -- License data
       CREATE TABLE IF NOT EXISTS license_data (
         node_id         uuid,
-        req_cnt         integer,
+        req_cnt         bigint,
         PRIMARY KEY (node_id)
       );
 
@@ -166,7 +166,7 @@ return {
         req_cnt         counter,
         PRIMARY KEY (node_id)
       );
-    
+
       /* Login Attempts */
       CREATE TABLE IF NOT EXISTS login_attempts (
         consumer_id uuid,
